+++ src/Stripe.net/Entities/Issuing/Cards/Card.cs
@@ -29,10 +29,10 @@ namespace Stripe.Issuing
         public DateTime Created { get; set; }
 
         [JsonProperty("exp_month")]
-        public int ExpMonth { get; set; }
+        public long ExpMonth { get; set; }
 
         [JsonProperty("exp_year")]
-        public int ExpYear { get; set; }
+        public long ExpYear { get; set; }
 
         [JsonProperty("last4")]
         public string Last4 { get; set; }
