+++ src/Stripe.net/Entities/Issuing/Cards/CardDetails.cs
@@ -33,10 +33,10 @@ namespace Stripe.Issuing
         public string Cvc { get; set; }
 
         [JsonProperty("exp_month")]
-        public int ExpMonth { get; set; }
+        public long ExpMonth { get; set; }
 
         [JsonProperty("exp_year")]
-        public int ExpYear { get; set; }
+        public long ExpYear { get; set; }
 
         [JsonProperty("number")]
         public string Number { get; set; }
